// src/types/series.d.ts
type ISerie = {
  adult: boolean;
  backdrop_path: string;
  first_air_date: string;
  genre_ids: Array<number>;
  id: number;
  origin_country: Array<string>;
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  name: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};

type ISerieDetails = {
  adult: boolean;
  backdrop_path: string;
  belongs_to_collection: {
    id: number;
    name: string;
    poster_path: string;
    backdrop_path: string;
  };
  budget: number;
  genres: [
    { id: number; name: string },
    { id: number; name: string },
    { id: number; name: string },
    { id: number; name: string },
    { id: number; name: string },
    { id: number; name: string }
  ];
  homepage: string;
  id: number;
  imdb_id: string;
  original_language: string;
  original_name: string;
  overview: string;
  popularity: number;
  poster_path: string;
  production_companies: Array<{
    id: number;
    logo_path: string;
    name: string;
    origin_country: string;
  }>;
  production_countries: Array<{ iso_3166_1: string; name: string }>;
  release_date: string;
  revenue: number;
  runtime: number;
  spoken_languages: Array<{
    english_name: string;
    iso_639_1: string;
    name: string;
  }>;
  status: string;
  tagline: string;
  name: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};

type ISeries = Array<ISerie>;

export { ISerie, ISeries, ISerieDetails };